requirements.txt
langchain
langchain-community
langchain-experimental
langchain-core
langchain-text-splitters==0.2.2
langchain-openai
pypdf
faiss-cpu
langchainhub
langsmith
langgraph
langchain-pinecone
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4
altair<5
streamlit
qrcode
pandas
tabulate
pygame